restructure CondTools/SiPixel
