+++ CondTools/SiPixel/plugins/SiPixelCondObjAllPayloadsReader.cc
@@ -1,14 +1,76 @@
+// -*- C++ -*-
+//
+// Package:    SiPixelCondObjAllPayloadsReader
+// Class:      SiPixelCondObjAllPayloadsReader
+//
+/**\class SiPixelCondObjAllPayloadsReader SiPixelCondObjAllPayloadsReader.h SiPixel/test/SiPixelCondObjAllPayloadsReader.h
+
+ Description: Test analyzer for reading pixel calibration from the DB
+
+ Implementation:
+     <Notes on implementation>
+*/
+//
+// Original Author:  Vincenzo CHIOCHIA
+//         Created:  Tue Oct 17 17:40:56 CEST 2006
+// $Id: SiPixelCondObjAllPayloadsReader.h,v 1.4 2009/05/28 22:12:54 dlange Exp $
+//
+//
+
+// system includes
 #include <memory>
 
-#include "SiPixelCondObjAllPayloadsReader.h"
+// user includes
+#include "CalibTracker/SiPixelESProducers/interface/SiPixelGainCalibrationForHLTService.h"
+#include "CalibTracker/SiPixelESProducers/interface/SiPixelGainCalibrationOfflineService.h"
+#include "CalibTracker/SiPixelESProducers/interface/SiPixelGainCalibrationService.h"
+#include "CalibTracker/SiPixelESProducers/interface/SiPixelGainCalibrationServiceBase.h"
+#include "CommonTools/UtilAlgos/interface/TFileService.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
 #include "Geometry/CommonDetUnit/interface/PixelGeomDetUnit.h"
 #include "Geometry/CommonTopologies/interface/PixelTopology.h"
-#include "FWCore/ServiceRegistry/interface/Service.h"
-#include "CommonTools/UtilAlgos/interface/TFileService.h"
+#include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
+#include "Geometry/TrackerGeometryBuilder/interface/TrackerGeometry.h"
+
+// ROOT includes
+#include "TROOT.h"
+#include "TFile.h"
+#include "TTree.h"
+#include "TBranch.h"
+#include "TH1F.h"
+#include <string>
+
+namespace cms {
+  class SiPixelCondObjAllPayloadsReader : public edm::one::EDAnalyzer<edm::one::SharedResources> {
+  public:
+    explicit SiPixelCondObjAllPayloadsReader(const edm::ParameterSet& iConfig);
+
+    void analyze(const edm::Event&, const edm::EventSetup&) override;
+    void endJob() override;
+
+  private:
+    const edm::ESGetToken<TrackerGeometry, TrackerDigiGeometryRecord> tkGeomToken_;
+    std::unique_ptr<SiPixelGainCalibrationServiceBase> SiPixelGainCalibrationService_;
+
+    std::map<uint32_t, TH1F*> _TH1F_Pedestals_m;
+    std::map<uint32_t, TH1F*> _TH1F_Gains_m;
+    TH1F* _TH1F_Gains_sum;
+    TH1F* _TH1F_Pedestals_sum;
+    TH1F* _TH1F_Gains_all;
+    TH1F* _TH1F_Pedestals_all;
+  };
+}  // namespace cms
 
 namespace cms {
   SiPixelCondObjAllPayloadsReader::SiPixelCondObjAllPayloadsReader(const edm::ParameterSet& conf)
-      : conf_(conf), tkGeomToken_(esConsumes()) {
+      : tkGeomToken_(esConsumes()) {
     usesResource(TFileService::kSharedResource);
     std::string payloadType = conf.getParameter<std::string>("payloadType");
     if (strcmp(payloadType.c_str(), "HLT") == 0) {
