+++ CondTools/SiPixel/plugins/SiPixelCondObjForHLTReader.cc
@@ -1,17 +1,87 @@
+// -*- C++ -*-
+//
+// Package:    SiPixelCondObjForHLTReader
+// Class:      SiPixelCondObjForHLTReader
+//
+/**\class SiPixelCondObjForHLTReader SiPixelCondObjForHLTReader.h SiPixel/plugins/SiPixelCondObjForHLTReader.cc
+
+ Description: Test analyzer for reading pixel calibration from the DB
+
+ Implementation:
+     <Notes on implementation>
+*/
+//
+// Original Author:  Vincenzo CHIOCHIA
+//         Created:  Tue Oct 17 17:40:56 CEST 2006
+// $Id: SiPixelCondObjForHLTReader.h,v 1.7 2009/05/28 22:12:55 dlange Exp $
+//
+//
+
+// system includes
 #include <memory>
 
-#include "SiPixelCondObjForHLTReader.h"
-
-#include "Geometry/CommonDetUnit/interface/PixelGeomDetUnit.h"
-#include "Geometry/CommonTopologies/interface/PixelTopology.h"
-#include "FWCore/ServiceRegistry/interface/Service.h"
-#include "CommonTools/UtilAlgos/interface/TFileService.h"
-#include "DataFormats/SiPixelDetId/interface/PixelSubdetector.h"
+// user includes
 #include "CalibTracker/SiPixelESProducers/interface/SiPixelGainCalibrationForHLTService.h"
 #include "CalibTracker/SiPixelESProducers/interface/SiPixelGainCalibrationForHLTSimService.h"
+#include "CalibTracker/SiPixelESProducers/interface/SiPixelGainCalibrationServiceBase.h"
+#include "CommonTools/UtilAlgos/interface/TFileService.h"
+#include "DataFormats/SiPixelDetId/interface/PixelSubdetector.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
+#include "Geometry/CommonDetUnit/interface/PixelGeomDetUnit.h"
+#include "Geometry/CommonTopologies/interface/PixelTopology.h"
+#include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
+#include "Geometry/TrackerGeometryBuilder/interface/TrackerGeometry.h"
+
+// ROOT includes
+#include "TROOT.h"
+#include "TFile.h"
+#include "TTree.h"
+#include "TBranch.h"
+#include "TH1F.h"
 
 namespace cms {
-  SiPixelCondObjForHLTReader::SiPixelCondObjForHLTReader(const edm::ParameterSet& conf)
+  class SiPixelCondObjForHLTReader : public edm::one::EDAnalyzer<edm::one::SharedResources> {
+  public:
+    explicit SiPixelCondObjForHLTReader(const edm::ParameterSet &iConfig);
+
+    void analyze(const edm::Event &, const edm::EventSetup &) override;
+    void endJob() override;
+
+  private:
+    edm::ParameterSet conf_;
+    //edm::ESHandle<SiPixelGainCalibration> SiPixelGainCalibration_;
+    const edm::ESGetToken<TrackerGeometry, TrackerDigiGeometryRecord> tkGeomToken_;
+    std::unique_ptr<SiPixelGainCalibrationServiceBase> SiPixelGainCalibrationService_;
+
+    std::map<uint32_t, TH1F *> _TH1F_Pedestals_m;
+    std::map<uint32_t, TH1F *> _TH1F_Gains_m;
+    std::map<uint32_t, double> _deadfrac_m;
+    std::map<uint32_t, double> _noisyfrac_m;
+
+    TH1F *_TH1F_Dead_sum;
+    TH1F *_TH1F_Noisy_sum;
+    TH1F *_TH1F_Gains_sum;
+    TH1F *_TH1F_Pedestals_sum;
+    TH1F *_TH1F_Dead_all;
+    TH1F *_TH1F_Noisy_all;
+    TH1F *_TH1F_Gains_all;
+    TH1F *_TH1F_Pedestals_all;
+    TH1F *_TH1F_Gains_bpix;
+    TH1F *_TH1F_Gains_fpix;
+    TH1F *_TH1F_Pedestals_bpix;
+    TH1F *_TH1F_Pedestals_fpix;
+  };
+}  // namespace cms
+
+namespace cms {
+  SiPixelCondObjForHLTReader::SiPixelCondObjForHLTReader(const edm::ParameterSet &conf)
       : conf_(conf), tkGeomToken_(esConsumes()) {
     if (conf_.getParameter<bool>("useSimRcd"))
       SiPixelGainCalibrationService_ =
@@ -21,7 +91,7 @@ namespace cms {
           std::make_unique<SiPixelGainCalibrationForHLTService>(conf_, consumesCollector());
   }
 
-  void SiPixelCondObjForHLTReader::analyze(const edm::Event& iEvent, const edm::EventSetup& iSetup) {
+  void SiPixelCondObjForHLTReader::analyze(const edm::Event &iEvent, const edm::EventSetup &iSetup) {
     //Create Subdirectories
     edm::Service<TFileService> fs;
     TFileDirectory subDirPed = fs->mkdir("Pedestals");
@@ -39,7 +109,7 @@ namespace cms {
         << "[SiPixelCondObjForHLTReader::beginJob] End Reading CondObjForHLTects" << std::endl;
 
     // Get the Geometry
-    const TrackerGeometry* tkgeom = &iSetup.getData(tkGeomToken_);
+    const TrackerGeometry *tkgeom = &iSetup.getData(tkGeomToken_);
     edm::LogInfo("SiPixelCondObjForHLTReader") << " There are " << tkgeom->dets().size() << " detectors" << std::endl;
 
     // Get the list of DetId's
@@ -70,7 +140,7 @@ namespace cms {
     _TH1F_Gains_bpix = fs->make<TH1F>("GainsBpix", "bpix Gains", 100, 0, 10);
     _TH1F_Gains_fpix = fs->make<TH1F>("GainsFpix", "fpix Gains", 100, 0, 10);
 
-    TTree* tree = new TTree("tree", "tree");
+    TTree *tree = new TTree("tree", "tree");
     uint32_t detid;
     double gainmeanfortree, gainrmsfortree, pedmeanfortree, pedrmsfortree;
     tree->Branch("detid", &detid, "detid/I");
@@ -91,8 +161,8 @@ namespace cms {
       _TH1F_Gains_m[detid] = subDirGain.make<TH1F>(name, name, 100, 0., 10.);
 
       DetId detIdObject(detid);
-      const PixelGeomDetUnit* _PixelGeomDetUnit =
-          dynamic_cast<const PixelGeomDetUnit*>(tkgeom->idToDetUnit(DetId(detid)));
+      const PixelGeomDetUnit *_PixelGeomDetUnit =
+          dynamic_cast<const PixelGeomDetUnit *>(tkgeom->idToDetUnit(DetId(detid)));
       if (_PixelGeomDetUnit == nullptr) {
         edm::LogError("SiPixelCondObjHLTDisplay") << "[SiPixelCondObjHLTReader::beginJob] the detID " << detid
                                                   << " doesn't seem to belong to Tracker" << std::endl;
@@ -104,9 +174,9 @@ namespace cms {
 
       nmodules++;
 
-      const GeomDetUnit* geoUnit = tkgeom->idToDetUnit(detIdObject);
-      const PixelGeomDetUnit* pixDet = dynamic_cast<const PixelGeomDetUnit*>(geoUnit);
-      const PixelTopology& topol = pixDet->specificTopology();
+      const GeomDetUnit *geoUnit = tkgeom->idToDetUnit(detIdObject);
+      const PixelGeomDetUnit *pixDet = dynamic_cast<const PixelGeomDetUnit *>(geoUnit);
+      const PixelTopology &topol = pixDet->specificTopology();
 
       // Get the module sizes.
       int nrows = topol.nrows();     // rows in x
