+++ CondTools/SiPixel/plugins/SiPixelLorentzAngleReader.cc
@@ -1,19 +1,51 @@
-#include "SiPixelLorentzAngleReader.h"
-#include "Geometry/CommonDetUnit/interface/PixelGeomDetUnit.h"
-#include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
-#include "DataFormats/SiPixelDetId/interface/PixelSubdetector.h"
-#include "DataFormats/DetId/interface/DetId.h"
-#include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
-#include "Geometry/CommonDetUnit/interface/PixelGeomDetUnit.h"
-#include "FWCore/ServiceRegistry/interface/Service.h"
-#include "CommonTools/UtilAlgos/interface/TFileService.h"
-#include "DataFormats/TrackerCommon/interface/PixelBarrelName.h"
-#include "DataFormats/TrackerCommon/interface/PixelEndcapName.h"
-#include "DataFormats/SiPixelDetId/interface/PixelSubdetector.h"
-
+// system include files
 #include <cstdio>
 #include <iostream>
 #include <sys/time.h>
+
+// user include files
+#include "CommonTools/UtilAlgos/interface/TFileService.h"
+#include "CondFormats/DataRecord/interface/SiPixelLorentzAngleRcd.h"
+#include "CondFormats/DataRecord/interface/SiPixelLorentzAngleSimRcd.h"
+#include "CondFormats/SiPixelObjects/interface/SiPixelLorentzAngle.h"
+#include "DataFormats/DetId/interface/DetId.h"
+#include "DataFormats/SiPixelDetId/interface/PixelSubdetector.h"
+#include "DataFormats/TrackerCommon/interface/PixelBarrelName.h"
+#include "DataFormats/TrackerCommon/interface/PixelEndcapName.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
+#include "Geometry/CommonDetUnit/interface/PixelGeomDetUnit.h"
+#include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
+
+// ROOT includes
+#include "TROOT.h"
+#include "TFile.h"
+#include "TH2F.h"
+
+//
+//
+// class decleration
+//
+class SiPixelLorentzAngleReader : public edm::one::EDAnalyzer<edm::one::SharedResources> {
+public:
+  explicit SiPixelLorentzAngleReader(const edm::ParameterSet&);
+  ~SiPixelLorentzAngleReader() override;
+
+  void analyze(const edm::Event&, const edm::EventSetup&) override;
+
+private:
+  const edm::ESGetToken<SiPixelLorentzAngle, SiPixelLorentzAngleRcd> siPixelLAToken_;
+  const edm::ESGetToken<SiPixelLorentzAngle, SiPixelLorentzAngleSimRcd> siPixelSimLAToken_;
+  const bool printdebug_;
+  const bool useSimRcd_;
+
+  TH1F* LorentzAngleBarrel_;
+  TH1F* LorentzAngleForward_;
+};
 
 using namespace cms;
 
@@ -53,4 +85,7 @@ void SiPixelLorentzAngleReader::analyze(const edm::Event& e, const edm::EventSet
     }
   }
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
 DEFINE_FWK_MODULE(SiPixelLorentzAngleReader);
