+++ CondTools/SiPixel/plugins/SiPixelPerformanceSummaryReader.cc
@@ -1,12 +1,36 @@
-#include "SiPixelPerformanceSummaryReader.h"
-#include "FWCore/MessageLogger/interface/MessageLogger.h"
-
+// system includes
 #include <cstdio>
 #include <iostream>
 #include <sys/time.h>
 
+// user includes
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "CondFormats/SiPixelObjects/interface/SiPixelPerformanceSummary.h"
+#include "CondFormats/DataRecord/interface/SiPixelPerformanceSummaryRcd.h"
+
 using namespace cms;
 using namespace std;
+
+namespace cms {
+  class SiPixelPerformanceSummaryReader : public edm::one::EDAnalyzer<> {
+  public:
+    explicit SiPixelPerformanceSummaryReader(const edm::ParameterSet&);
+    ~SiPixelPerformanceSummaryReader() override;
+
+    void analyze(const edm::Event&, const edm::EventSetup&) override;
+
+  private:
+    const edm::ESGetToken<SiPixelPerformanceSummary, SiPixelPerformanceSummaryRcd> perfSummaryToken_;
+    const bool printdebug_;
+  };
+}  // namespace cms
 
 SiPixelPerformanceSummaryReader::SiPixelPerformanceSummaryReader(const edm::ParameterSet& iConfig)
     : perfSummaryToken_(esConsumes()), printdebug_(iConfig.getUntrackedParameter<bool>("printDebug", false)) {}
